Switch from threads to async tasks (#46)
* Redefine `__eq__` special method for `SunServer` and `SunUser` classes

This commit add redefinition of the special method `__eq__` for both
`SunServer` and `SunUser` Sunbot classes. Two instances of these
classes are equal if and only if they have the same Discord ID.

* Add weather_event module

* Create daily weather task in corresponding class

This new task will replace corresponding thread in `Meteo` module in a
futur commit.

* Link new daily weather task into `SunController`

* Fix runtime issue

There was an error in import part

* Add `daily_weather` slash command

This new command allows to indicate to the bot the channel where
we want that the bot sends the daily weather image. It is possible
to disable the sending or to change the channel on the server with
the same command.

* Fix some issues in `weather_event` module

* Add docstring to `weather_event` module

* Remove `server_id` from the arguments of `add_srv2location` method

Server ID is a data that is contained in discord `interaction`
which is also passed in parameters of this method

* Update daily weather hour sending

* Remove unused import in `weather_event` module